+++ juriscraper/opinions/united_states/state/ind.py
@@ -6,6 +6,7 @@ Auth: Jon Andersen <janderse@gmail.com>
 Reviewer: mlr
 History:
     2014-09-03: Created by Jon Andersen
+    2025-07-09: Updated bt luism add new fields for lower court details and judge names
 """
 
 from juriscraper.OpinionSiteLinear import OpinionSiteLinear
@@ -30,5 +31,19 @@ class Site(OpinionSiteLinear):
                     "date": case["date"],
                     "url": f"https://public.courts.in.gov/Decisions/{case['opinionUrl']}",
                     "disposition": case["decision"],
+                    "lower_court": ", ".join(
+                        court["name"]
+                        for court in case["courts"]
+                        if court["name"] != self.court_name
+                    ),
+                    "lower_court_number": ", ".join(
+                        court["number"]
+                        for court in case["courts"]
+                        if court["name"] != self.court_name
+                    ),
+                    "judge": ", ".join(
+                        vote["judge"].split(",")[-1].replace(".", "")
+                        for vote in case["opinion"]["votes"]
+                    ),
                 }
             )